mysql_connector.py
import pymysql
import os
from dotenv import load_dotenv
load_dotenv()

config = {
    'host': os.getenv('MYSQL_HOST'),
    'user': os.getenv('MYSQL_USER'),
    'password': os.getenv('MYSQL_PASSWORD'),
    'database': os.getenv('MYSQL_DATABASE')
}

connection = pymysql.connect(**config)
