requirements.txt
dnspython==2.8.0
pymongo==4.16.0
PyMySQL==1.1.2
python-dotenv==1.2.2
tabulate
